requirements.txt
fastapi>=0.100.0,<1
uvicorn[standard]>=0.20.0,<1
psycopg2-binary>=2.8.0,<3
pydantic>=2.0.0,<3
python-dotenv>=0.19.0,<2
click>=7.0,<9